High DPI hack
This fixes a problem that shows up with gloss programs on my MacBook
Pro's Retina screen: only the lower-left quadrant of the window gets
used; the rest remains blank.

This problem hasn't always happened. I think some MacOS update must have
changed OpenGL or GLUT in a way that introduced the problem (probably
10.15). In any case, the problem can be fixed by setting the GL viewport
to be double the GLUT window size in both dimensions.